used HarDataEntry for har generation
